put azure in seperate config file
